+++ clients/_b2bi/client.go
@@ -14,7 +14,7 @@ package _b2bi
 import (
 	"sync"
 
-	"github.com/TouchBistro/aws-ccp-go/providers"
+	"github.com/TouchBistro/awesome/providers"
 	"github.com/aws/aws-sdk-go-v2/service/b2bi"
 )