+++ clients/_s3control/client.go
@@ -14,7 +14,7 @@ package _s3control
 import (
 	"sync"
 
-	"github.com/TouchBistro/aws-ccp-go/providers"
+	"github.com/TouchBistro/awesome/providers"
 	"github.com/aws/aws-sdk-go-v2/service/s3control"
 )