+++ clients/_sagemakera2iruntime/client.go
@@ -14,7 +14,7 @@ package _sagemakera2iruntime
 import (
 	"sync"
 
-	"github.com/TouchBistro/aws-ccp-go/providers"
+	"github.com/TouchBistro/awesome/providers"
 	"github.com/aws/aws-sdk-go-v2/service/sagemakera2iruntime"
 )