+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 
 
 ## Version v0.0.5
-- Import github.com/Touchbistro/aws-ccp-go/init/cmd with a blank import & get command-line `default` provider setup [check this link](./init/cmd/README.md).
+- Import github.com/TouchBistro/awesome/init/cmd with a blank import & get command-line `default` provider setup [check this link](./init/cmd/README.md).
 - `MustDefault()` function added to `providers` which returns the `default` named provider or panics